+++ test/testdata/demopackage/__init__.py
@@ -7,6 +7,8 @@ from . import child_b
 from ._child_d import Test
 from .child_b import B
 from .child_c import C
+from .child_f import F
+from .subpackage import G
 
 if demopackage2:
     pass
@@ -17,8 +19,11 @@ __all__ = [
     "C",
     "child_b",
     "child_c",
+    "child_f",
     "demopackage2",
     "_child_e",
     "child_excluded",
     "subpackage",
+    "F",
+    "G",
 ]