+++ test/testdata/demopackage/child_c.py
@@ -4,7 +4,7 @@ from . import child_b
 
 
 class C(child_b.B):
-    """This class is defined in .child_c and inherits from .child_b.B"""
+    """This class is defined in .child_c and inherits from .child_b.B."""
 
     def c(self):
         return 2