+++ test/testdata/demopackage/subpackage/child_g.py
@@ -1,0 +1,8 @@
+"""child_g docstring"""
+
+
+class G:
+    """This class is defined in subpackage's child_g, but reexposed in the `demopackage.subpackage` namespace."""
+
+    def g(self):
+        pass